FIX: Renames mail_sender to correct convention and fix assignment alignment
